+++ engine/runtime/src/main/scala/org/enso/compiler/codegen/IrToTruffle.scala
@@ -5,7 +5,13 @@ import com.oracle.truffle.api.frame.FrameSlot
 import com.oracle.truffle.api.source.{Source, SourceSection}
 import org.enso.compiler.core.IR
 import org.enso.compiler.core.IR.Module.Scope.Import
+import org.enso.compiler.core.IR.Name.Special
 import org.enso.compiler.core.IR.{Error, IdentifiedLocation, Pattern}
+import org.enso.compiler.data.BindingsMap.{
+  ExportedModule,
+  ResolvedConstructor,
+  ResolvedModule
+}
 import org.enso.compiler.data.{BindingsMap, CompilerConfig}
 import org.enso.compiler.exception.{BadPatternMatch, CompilerError}
 import org.enso.compiler.pass.analyse.AliasAnalysis.Graph.{Scope => AliasScope}
@@ -51,20 +57,20 @@ import org.enso.interpreter.node.{
   ExpressionNode => RuntimeExpression
 }
 import org.enso.interpreter.runtime.Context
-import org.enso.interpreter.runtime.callable.{
-  UnresolvedConversion,
-  UnresolvedSymbol
-}
 import org.enso.interpreter.runtime.callable.argument.{
   ArgumentDefinition,
   CallArgument
 }
-import org.enso.interpreter.runtime.callable.atom.Atom
-import org.enso.interpreter.runtime.callable.atom.AtomConstructor
+import org.enso.interpreter.runtime.callable.atom.{Atom, AtomConstructor}
 import org.enso.interpreter.runtime.callable.function.{
   FunctionSchema,
   Function => RuntimeFunction
 }
+import org.enso.interpreter.runtime.callable.{
+  UnresolvedConversion,
+  UnresolvedSymbol
+}
+import org.enso.interpreter.runtime.data.Type
 import org.enso.interpreter.runtime.data.text.Text
 import org.enso.interpreter.runtime.scope.{
   FramePointer,
@@ -74,13 +80,9 @@ import org.enso.interpreter.runtime.scope.{
 import org.enso.interpreter.{Constants, Language}
 
 import java.math.BigInteger
-import org.enso.compiler.core.IR.Name.Special
-import org.enso.interpreter.runtime.data.Type
-
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 import scala.jdk.OptionConverters._
-import scala.jdk.CollectionConverters._
 
 /** This is an implementation of a codegeneration pass that lowers the Enso
   * [[IR]] into the truffle [[org.enso.compiler.core.Core.Node]] structures that
@@ -152,18 +154,31 @@ class IrToTruffle(
   private def processModule(module: IR.Module): Unit = {
     generateMethods()
     generateReExportBindings(module)
-    module
-      .unsafeGetMetadata(
-        BindingAnalysis,
-        "No binding analysis at the point of codegen."
-      )
-      .resolvedExports
+    val bindingsMap =
+      module
+        .unsafeGetMetadata(
+          BindingAnalysis,
+          "No binding analysis at the point of codegen."
+        )
+
+    bindingsMap.resolvedExports
+      .collect { case ExportedModule(ResolvedModule(module), _, _) =>
+        module
+      }
       .foreach { exp =>
-        moduleScope.addExport(exp.module.unsafeAsModule().getScope)
+        moduleScope.addExport(exp.unsafeAsModule().getScope)
       }
     val imports = module.imports
     val methodDefs = module.bindings.collect {
       case method: IR.Module.Scope.Definition.Method.Explicit => method
+    }
+
+    bindingsMap.resolvedImports.foreach { imp =>
+      imp.target match {
+        case BindingsMap.ResolvedType(_, _) =>
+        case ResolvedModule(module) =>
+          moduleScope.addImport(module.unsafeAsModule().getScope)
+      }
     }
 
     // Register the imports in scope
@@ -173,11 +188,8 @@ class IrToTruffle(
           poly.getVisibleName,
           context.getEnvironment.lookupHostSymbol(i.getJavaName)
         )
-      case i: Import.Module =>
-        this.moduleScope.addImport(
-          context.getCompiler.processImport(i.name.name, i.location, source)
-        )
-      case _: Error =>
+      case _: Import.Module =>
+      case _: Error         =>
     }
 
     val typeDefs = module.bindings.collect {
@@ -186,9 +198,10 @@ class IrToTruffle(
 
     typeDefs.foreach { tpDef =>
       // Register the atoms and their constructors in scope
-      val atomDefs = tpDef.members
+      val atomDefs    = tpDef.members
+      val runtimeType = moduleScope.getTypes.get(tpDef.name.name)
       val atomConstructors =
-        atomDefs.map(cons => moduleScope.getConstructors.get(cons.name.name))
+        atomDefs.map(cons => runtimeType.getConstructors.get(cons.name.name))
       atomConstructors
         .zip(atomDefs)
         .foreach { case (atomCons, atomDefn) =>
@@ -247,8 +260,7 @@ class IrToTruffle(
             )
           }
         }
-      val tp = moduleScope.getTypes.get(tpDef.name.name)
-      tp.generateGetters(language, atomConstructors.asJava)
+      runtimeType.generateGetters(language)
     }
 
     // Register the method definitions in scope
@@ -265,7 +277,7 @@ class IrToTruffle(
         "Method definition missing dataflow information."
       )
 
-      val consOpt =
+      val declaredConsOpt =
         methodDef.methodReference.typePointer match {
           case None =>
             Some(moduleScope.getAssociatedType)
@@ -293,6 +305,12 @@ class IrToTruffle(
                 }
               }
         }
+
+      val consOpt = declaredConsOpt.map { c =>
+        if (methodDef.isStatic) {
+          c.getEigentype
+        } else { c }
+      }
 
       consOpt.foreach { cons =>
         val expressionProcessor = new ExpressionProcessor(
@@ -607,7 +625,10 @@ class IrToTruffle(
     )
     bindingsMap.exportedSymbols.foreach {
       case (name, resolution :: _) =>
-        if (resolution.module.unsafeAsModule() != moduleScope.getModule) {
+        if (
+          resolution.isInstanceOf[ResolvedConstructor] || resolution.module
+            .unsafeAsModule() != moduleScope.getModule
+        ) {
           resolution match {
             case BindingsMap.ResolvedType(module, tp) =>
               val runtimeTp =
@@ -618,10 +639,9 @@ class IrToTruffle(
                 name,
                 fun
               )
-            case BindingsMap.ResolvedConstructor(definitionModule, cons) =>
-              val runtimeCons = definitionModule
-                .unsafeAsModule()
-                .getScope
+            case BindingsMap.ResolvedConstructor(definitionType, cons) =>
+              val runtimeCons = definitionType
+                .unsafeToRuntimeType()
                 .getConstructors
                 .get(cons.name)
               val fun = mkConsGetter(runtimeCons)
@@ -927,11 +947,11 @@ class IrToTruffle(
                   )
                 case Some(
                       BindingsMap.Resolution(
-                        BindingsMap.ResolvedConstructor(mod, cons)
+                        BindingsMap.ResolvedConstructor(tp, cons)
                       )
                     ) =>
                   val atomCons =
-                    mod.unsafeAsModule().getScope.getConstructors.get(cons.name)
+                    tp.unsafeToRuntimeType().getConstructors.get(cons.name)
                   val r = if (atomCons == context.getBuiltins.bool().getTrue) {
                     BooleanBranchNode.build(true, branchCodeNode.getCallTarget)
                   } else if (atomCons == context.getBuiltins.bool().getFalse) {
@@ -1192,10 +1212,9 @@ class IrToTruffle(
                 ConstantObjectNode.build(
                   tp.module.unsafeAsModule().getScope.getTypes.get(tp.tp.name)
                 )
-              case BindingsMap.ResolvedConstructor(definitionModule, cons) =>
-                val c = definitionModule
-                  .unsafeAsModule()
-                  .getScope
+              case BindingsMap.ResolvedConstructor(definitionType, cons) =>
+                val c = definitionType
+                  .unsafeToRuntimeType()
                   .getConstructors
                   .get(cons.name)
                 if (c == null) {