+++ CHANGELOG.md
@@ -371,6 +371,7 @@
 - [main = "Hello World!" is valid Enso sample][3696]
 - [Invalidate module's IR cache if imported module changed][3703]
 - [Don't rename imported Main module that only imports names][3710]
+- [Distinguish static and instance methods][3740]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -419,6 +420,7 @@
 [3696]: https://github.com/enso-org/enso/pull/3696
 [3696]: https://github.com/enso-org/enso/pull/3703
 [3696]: https://github.com/enso-org/enso/pull/3710
+[3740]: https://github.com/enso-org/enso/pull/3740
 
 # Enso 2.0.0-alpha.18 (2021-10-12)